backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
httpx==0.28.1
python-dotenv==1.1.0
pydantic==2.11.3
python-multipart==0.0.20
sse-starlette==1.8.2
hypothesis==6.135.1
pytest==8.4.1
pytest-asyncio==1.0.0
